+++ packages/flutter_tools/test/general.shard/build_system/targets/localizations_test.dart
@@ -8,6 +8,7 @@ import 'package:flutter_tools/src/base/logger.dart';
 import 'package:flutter_tools/src/build_system/build_system.dart';
 import 'package:flutter_tools/src/build_system/targets/localizations.dart';
 import 'package:flutter_tools/src/localizations/gen_l10n.dart';
+import 'package:flutter_tools/src/localizations/localizations_utils.dart';
 import 'package:mockito/mockito.dart';
 
 import '../../../src/common.dart';
@@ -58,7 +59,7 @@ void main() {
         templateArbFileName: 'example.arb',
         outputFileString: 'bar',
         classNameString: 'Foo',
-        preferredSupportedLocale: <String>['en_US'],
+        preferredSupportedLocales: <String>['en_US'],
         headerString: 'HEADER',
         headerFile: 'header',
         useDeferredLoading: true,
