+++ packages/flutter_tools/test/general.shard/generate_localizations_test.dart
@@ -984,7 +984,7 @@ void main() {
           templateArbFileName: defaultTemplateArbFileName,
           outputFileString: defaultOutputFileString,
           classNameString: defaultClassNameString,
-          preferredSupportedLocale: preferredSupportedLocale,
+          preferredSupportedLocales: preferredSupportedLocale,
         );
         generator.loadResources();
       } on L10nException catch (e) {
@@ -1020,7 +1020,7 @@ void main() {
             templateArbFileName: defaultTemplateArbFileName,
             outputFileString: defaultOutputFileString,
             classNameString: defaultClassNameString,
-            preferredSupportedLocale: preferredSupportedLocale,
+            preferredSupportedLocales: preferredSupportedLocale,
           );
           generator.loadResources();
         } on L10nException catch (e) {
@@ -1433,7 +1433,7 @@ void main() {
           templateArbFileName: defaultTemplateArbFileName,
           outputFileString: defaultOutputFileString,
           classNameString: defaultClassNameString,
-          preferredSupportedLocale: preferredSupportedLocale,
+          preferredSupportedLocales: preferredSupportedLocale,
         );
         generator.loadResources();
         generator.writeOutputFiles(BufferLogger.test());
